fix: missing lodash-es, replace lodash-es to lodash
