+++ package.json
@@ -87,6 +87,7 @@
   "homepage": "https://github.com/bnb-chain/zkbnb-js-sdk#readme",
   "dependencies": {
     "@types/node": "^18.13.0",
+    "lodash": "^4.17.21",
     "shelljs": "^0.8.5"
   },
   "devDependencies": {
@@ -103,6 +104,7 @@
     "@semantic-release/changelog": "^6.0.1",
     "@semantic-release/git": "^10.0.1",
     "@types/chai": "^4.3.1",
+    "@types/lodash": "^4.14.192",
     "@types/mocha": "^9.1.1",
     "@typescript-eslint/eslint-plugin": "^5.32.0",
     "@typescript-eslint/parser": "^5.32.0",
