+++ src/query/index.ts
@@ -5,6 +5,5 @@ export * from './query';
 export * from './queryServer';
 export * from './response';
 export * from './type';
-export * from './errors';
 
 export const Query = typeof window === 'undefined' ? QueryServer : QueryClient;
