+++ src/query/response.ts
@@ -30,6 +30,22 @@ export class ResponseData<T = Data> extends Response {
   }
 }
 
+export class TimeoutError extends Error {}
+
+export class ResponseError extends Error {
+  public response: ResponseData;
+
+  public code: string | number;
+
+  constructor(message: string, response: ResponseData) {
+    super(message);
+    this.response = response;
+    this.code = response.status;
+  }
+}
+
 export type ResponseReturn = Promise<ResponseData> | ResponseData;
 
 export type ResponseHook = (response: ResponseData) => ResponseReturn;
+
+export type ErrorHook = (error: ResponseError) => ResponseReturn;
